server/requirements.txt
flask
flask_sqlalchemy
paho-mqtt
flask-socketio
psycopg2-binary
sqlalchemy